move pipeline decorator and infer short_name
