+++ garden_ai/container/containerize.py
@@ -6,7 +6,6 @@ from tempfile import TemporaryDirectory
 from threading import Thread
 from typing import List, Optional
 
-from garden_ai.pipelines import PipelineMetadata
 
 IMAGE_NAME = "gardenai/base"
 CONTAINER_NAME = "garden_ai"
@@ -126,18 +125,3 @@ def start_container(
         subprocess.run(["docker", "rmi", image_name])
 
     return output.stdout
-
-
-def garden_pipeline(metadata: PipelineMetadata, model_connectors=None):
-    if model_connectors is None:
-        model_connectors = []
-
-    def decorator(func):
-        def garden_target(*args, **kwargs):
-            return func(*args, **kwargs)
-
-        garden_target._pipeline_meta = metadata.dict()
-        garden_target._model_connectors = model_connectors
-        return garden_target  # returns func back, but with `__name__ == garden_target` and metadata
-
-    return decorator
